feat!: integrate diffusivity ROI analysis
**BREAKING**: ROI analysis is now performed on diffusivity measures

The ROI analysis step is generalizable to both fractional anisotropy and diffusivity analyses alike, and is thus moved in this commit to occur after diffusivity measures have been derived in order to execute ROI analysis on all measures (i.e., FA, MD, RD, & AD) at once as opposed to having to duplicate the step unnecessarily.

In order to do so, the step of cleaning up the file hierarchy after FA analysis was moved to before the diffusivity measure derivation step so as to standardize the file hierarchy before proceeding with ROI analysis.

This also necessitates changing the path of files used in other parts of the script, which have been modified appopriately.